src/plugins/pjsk_kits/__init__.py
from nonebot import get_plugin_config
from nonebot.plugin import PluginMetadata

__plugin_meta__ = PluginMetadata(
    name="pjsk_kits",
    description="",
    usage="",
)

from . import sheet as sheet
from .sticker import sticker as sticker