requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arclet-alconna==1.8.38
arclet-alconna-tools==0.7.10
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dashscope==1.23.6
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.14
frozenlist==1.7.0
fuzzywuzzy==0.18.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
loguru==0.7.3
msgpack==1.1.1
multidict==6.6.1
nepattern==0.7.7
nonebot-adapter-onebot==2.4.6
nonebot-plugin-alconna==0.58.5
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-waiter==0.8.1
nonebot2==2.4.2
openai==1.93.0
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pygtrie==2.5.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tarina==0.6.8
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.1
zipp==3.23.0
